Princeton poster as .xcf and NEIU flyer as .svg
Princeton poster was converted to RGB color space and all the text
layers removed, because they used commercial fonts. Xcf is the GIMP
native file format. NEIU flyer layout and graphical elements were
converted to .svg vector file format as an approximation and I'd be
happy to see other takes on how the ground is represented. The flyer has
the OpenHatch logo text as paths so Tuffy font doesn't have to be
installed.